% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_upload.R
\name{cl_raw_query}
\alias{cl_raw_query}
\title{根据品牌及开始日期结束日期查询相应的语料信息}
\usage{
cl_raw_query(brand, startDate, endDate)
}
\arguments{
\item{brand}{品牌}

\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
根据品牌及开始日期结束日期查询相应的语料信息
}
\examples{
cl_raw_query();
}
