% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_upload.R
\name{cl_raw_read}
\alias{cl_raw_read}
\title{读取原始的语料信息}
\usage{
cl_raw_read(files, brand)
}
\arguments{
\item{files}{文件}

\item{brand}{品牌}
}
\value{
返回值
}
\description{
读取原始的语料信息
}
\examples{
cl_raw_read();
}
