% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_upload.R
\name{cl_raw_upload_with_msg}
\alias{cl_raw_upload_with_msg}
\title{上传原始语料进入服务器}
\usage{
cl_raw_upload_with_msg(data)
}
\arguments{
\item{data}{数据}
}
\value{
返回值
}
\description{
上传原始语料进入服务器
}
\examples{
cl_raw_upload_with_msg();
}
