% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_aux_funcs.R
\name{is_new_cl_data}
\alias{is_new_cl_data}
\title{判断是否为最新的语料}
\usage{
is_new_cl_data(FBrand, FDate)
}
\arguments{
\item{FBrand}{品牌}

\item{FDate}{日期}
}
\value{
返回值
}
\description{
判断是否为最新的语料
}
\examples{
is_new_cl_data();
}
