% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qalist.R
\name{is_new_qalist}
\alias{is_new_qalist}
\title{判断是否为最新的QA对信息}
\usage{
is_new_qalist(FBrand, FDate)
}
\arguments{
\item{FBrand}{品牌}

\item{FDate}{日期}
}
\value{
返回值
}
\description{
判断是否为最新的QA对信息
}
\examples{
is_new_qalist();
}
