% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qalist.R
\name{qalist_gen}
\alias{qalist_gen}
\title{处理QA配对信息}
\usage{
qalist_gen(brand, startDate, endDate)
}
\arguments{
\item{brand}{品牌}

\item{startDate}{开始日期}

\item{endDate}{结束日期}
}
\value{
返回值
}
\description{
处理QA配对信息
}
\examples{
qalist_gen();
}
