% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_id_deal.R
\name{qalist_idize}
\alias{qalist_idize}
\title{按品牌处理问题答案及对话的ID化问题}
\usage{
qalist_idize(brand)
}
\arguments{
\item{brand}{品牌}
}
\value{
返回消息内容
}
\description{
按品牌处理问题答案及对话的ID化问题
}
\examples{
qalist_idize();
}
